Consolidate all Transcode tests in single file.
Addresses #58.
